Rename ParsingException to QueryParsingException
