+++ InformationScripting/src/parsing/QueryParsingException.cpp
